+++ src/api/api.ts
@@ -218,6 +218,7 @@ export default class MastoApi {
 
     private apiMutexes = simpleCacheKeyDict(() => new Mutex());   // For locking data fetching for an API endpoint
     private cacheMutexes = simpleCacheKeyDict(() => new Mutex()); // For locking checking the cache for an API endpoint
+    private isHomeserverGoToSocial: Optional<boolean> = undefined;
     private requestSemphore = new Semaphore(config.api.maxConcurrentHashtagRequests); // Concurrency of search & hashtag requests
 
     /**
@@ -405,7 +406,7 @@ export default class MastoApi {
             return toots;
         } catch (err) {
             // TODO: the hacky cast is because ApiCacheKey is broader than CacheKey
-            this.handleApiError({ cacheKey: cacheKey as CacheKey, logger }, [], err);
+            await this.handleApiError({ cacheKey: cacheKey as CacheKey, logger }, [], err);
             return [];
         } finally {
             this.waitTimes[cacheKey].markEnd();
@@ -695,6 +696,26 @@ export default class MastoApi {
     }
 
     /**
+     * Return true if the user's home server is a GoToSocial server.
+     * @returns {Promise<boolean>}
+     */
+    async isGoToSocialUser(): Promise<boolean> {
+        if (isNil(this.isHomeserverGoToSocial)) {
+            this.logger.debug(`Checking if user's home server is GoToSocial...`);
+            const instance = await this.instanceInfo();
+            this.isHomeserverGoToSocial = instance?.sourceUrl?.endsWith('gotosocial')
+
+            if (typeof this.isHomeserverGoToSocial !== 'boolean') {
+                this.logger.warn(`Failed to set isHomeserverGoToSocial to bool, sourceUrl: "${instance?.sourceUrl}", instance`, instance);
+            } else {
+                this.logger.debug(`Set isHomeserverGoToSocial to ${this.isHomeserverGoToSocial} based on instance:`, instance);
+            }
+        }
+
+        return !!this.isHomeserverGoToSocial;
+    }
+
+    /**
      * Locks all API and cache mutexes for cache state operations.
      * @returns {Promise<ConcurrencyLockRelease[]>} Array of lock release functions.
      */
@@ -865,7 +886,7 @@ export default class MastoApi {
             if (cacheKey != CacheKey.HASHTAG_TOOTS) logger.info(`Retrieved ${newRows.length} objects`);
             return newRows;
         } catch (e) {
-            return this.handleApiError<T>(params, newRows, e);
+            return await this.handleApiError<T>(params, newRows, e);
         }
     }
 
@@ -1092,17 +1113,28 @@ export default class MastoApi {
      * @param {Error | unknown} err - The error encountered.
      * @returns {T[]} Array of rows to use.
      */
-    private handleApiError<T extends ApiObj>(
+    private async handleApiError<T extends ApiObj>(
         params: Partial<FetchParamsWithCacheData<T>>,
         newRows: T[],
         err: Error | unknown,
-    ): ResponseRow<T>[] {
+    ): Promise<ResponseRow<T>[]> {
         const { cacheResult } = params;
         let { cacheKey, logger } = params;
         cacheKey ??= CacheKey.HOME_TIMELINE_TOOTS;  // TODO: this is a hack to avoid undefined cacheKey
         const waitTime = this.waitTimes[cacheKey];
+        const requestDefaults = config.api.data[cacheKey] ?? {};
         logger = logger ? logger.tempLogger('handleApiError') : getLogger(cacheKey, 'handleApiError');
+        logger.trace(`Handling API error for params:`, params, `\nerror:`, err);
         const cachedRows = cacheResult?.rows || [];
+
+        if (!newRows?.length && requestDefaults.canBeDisabledOnGoToSocial && await this.isGoToSocialUser()) {
+            const goToSocialWarning = config.api.errorMsgs.goToSocialHashtagTimeline(cacheKey);
+            const goToSocialError = logger.line(`Failed to fetch data. ${goToSocialWarning}`);
+            this.logger.warn(goToSocialError, err);
+            this.apiErrors.push(new Error(goToSocialError, {cause: err}));
+            return cachedRows;
+        }
+
         let msg = `"${err} after pulling ${newRows.length} rows (cache: ${cachedRows.length} rows).`;
         this.apiErrors.push(new Error(logger.line(msg), {cause: err}));
         throwIfAccessTokenRevoked(logger, err, `Failed ${waitTime.ageString()}. ${msg}`);