+++ src/index.ts
@@ -510,6 +510,14 @@ export default class TheAlgorithm {
     }
 
     /**
+     * True if fedialgo user is on a GoToSocial instance instead of plain vanilla Mastodon.
+     * @returns {boolean}
+     */
+    async isGoToSocialUser(): Promise<boolean> {
+        return await MastoApi.instance.isGoToSocialUser();
+    }
+
+    /**
      * Return info about the Fedialgo user's home mastodon instance.
      * @returns {Promise<mastodon.v2.Instance>} Instance info.
      */